Build instructions: add mroe qt packages for Debian and Ubuntu
I needed to install these packages for the build to work. Without them,
cmake failed looking for e.g. `Qt5LinguistToolsConfig.cmake`.

This has only been tested on my system, which is running Ubuntu 22.04.2
LTS.

Partly addresses
https://github.com/blueprint-freespeech/ricochet-refresh/issues/142,
though it looks like the reporter intends/intended to do a more thorough
analysis of the set of deps needed, so not sure it'd make sense to close
it.