build: Abseil as libtego dependency to fix linker errors
ricochet-refresh's build fails with a "DSO missing from command line"
error, on Arch Linux (binutils 2.42), when liking the final binary.
libabsl_log_internal_check_op.so is needed to build libtego.a because
protobuf depends on it. Setting the linker flags to
"-Wl,--copy-dt-needed-entries" works around the issue but goes against
the now default behaviour of binutils of not recursively fetching
dependencies for libs when linking. Adding this explicit link to Abseil
fixes the error in the proper manner but makes libtego build depend on
the internals of protobuf.